demux: hls: sched update from prev one
should avoid the issues with targetduration rounding
and inaccurate segment durations

(cherry picked from commit d443bd922c331070a211f075e2291ec96c5c2fb1)

+++ modules/demux/adaptive/playlist/BaseRepresentation.h
@@ -65,7 +65,7 @@ namespace adaptive
                 virtual void        pruneByPlaybackTime     (mtime_t);
 
                 virtual mtime_t     getMinAheadTime         (uint64_t) const;
-                virtual bool        needsUpdate             () const;
+                virtual bool        needsUpdate             (uint64_t) const;
                 virtual bool        runLocalUpdates         (SharedResources *);
                 virtual void        scheduleNextUpdate      (uint64_t, bool);
